feat: allow switching persistent store to session attributes instead (#181)
This change adds a new `useSessionAttributesForPersistentStore` property to the skill config at the root and deployment levels, that disables the use of the litexa persistent store backend and instead switches to using the skill session attributes for storage instead. This has the effect of changing persistent @variable lifespans from permanent to session, i.e. when this flag is set to true then all @ variables are deleted when the skill quits.