feat: new sessionTerminatingCallback for last change user code
This inserts a callback when the skill knows it'll terminate, either because it received `sessionEndedRequest` or is about to return `shouldEndSession:true`. Useful for e.g. reporting metrics, or cleaning up any resources that should share the same lifespan as a session.

note: the contents of the context object will vary depending on where this is called due to the skill being in a different state (e.g. no `context.say` when we're in `sessionEndedRequest`), but `context.request` will always have the request, and the `context.db` will be initialized with a working interface, letting the skill make any final adjustments before the session ends. This callback does NOT have access to the response though, as it can be called in situations where Litexa assumes there is only one valid response, e.g. at `sessionEndedRequest`.