feat: one shot intents are now delivered after launch is entered (#188)
This change primarily causes one shot intents (intents received during launch, e.g. "Alexa, ask [skill] to [intent]") to be delivered to the launch intent. Previously these could only be handled from the global state, and would be executed before the launch intent had been entered. This is implemented a new execution phase in the skill code that enters a state (currently always 'launch') before the existing intent handoff phase.

Other related changes Include:
* updates to documentation, new one shot test, and fixup of the hello world test to reflect the new oneshot behavior
* Litexa tests now also look at exit handler to match say statements
* exit handles are now correctly checked for statements that may introduce API declarations in the manifest, e.g. enabling APL
* exit handlers are now also allowed to specify state transitions, i.e. they can make a last minute audible and redirect flow to a different state
* subsequent statements in the group `-> END LISTEN` all now negate each other's effects. This lets you, e.g. undo the decision to END a skill based on a later signal. For example, you could have a default transition out of launch that gets turned into an END instead, given the presence of a particular one shot intent.
* the `request` statement now lets you define a test request of any type, loaded from a JSON file in the litexa directory. The previous functionality of defining an intent by type name was already handled by the `user` statement instead.
* there is a new litexa.sessionTerminatingCallback that is the last change user code to run when a session ends. Can be used to clean up state, or report session stats. See details below.


*litexa.sessionTerminatingCallback*
This inserts a callback when the skill knows it'll terminate, either because it received `sessionEndedRequest` or is about to return `shouldEndSession:true`. Useful for e.g. reporting metrics, or cleaning up any resources that should share the same lifespan as a session.

note: the contents of the context object will vary depending on where this is called due to the skill being in a different state (e.g. no `context.say` when we're in `sessionEndedRequest`), but `context.request` will always have the request, and the `context.db` will be initialized with a working interface, letting the skill make any final adjustments before the session ends. This callback does NOT have access to the response though, as it can be called in situations where Litexa assumes there is only one valid response, e.g. at `sessionEndedRequest`.