feat: scrubbing for non-inclusive language
Looking for instances of non-inclusive language, finding replacements for the terms 'whitelist' and 'blacklist'
Note: there was a user facing function call in the JSON validator tools that was called "whiteList" and has been renamed to "allowOnly". The old form has been left in for backwards compatibility but will emit a warning, and should no longer be used by new code.